[refactor🔨] : 띄어쓰기 리팩토링
