+++ src/main/java/shop/gaship/gashipshoppingmall/addresslist/service/Impl/AddressListServiceImpl.java
@@ -1,4 +1,4 @@
-package shop.gaship.gashipshoppingmall.addresslist.service;
+package shop.gaship.gashipshoppingmall.addresslist.service.Impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -15,6 +15,7 @@ import shop.gaship.gashipshoppingmall.addresslist.dto.AddressListResponseDto;
 import shop.gaship.gashipshoppingmall.addresslist.entity.AddressList;
 import shop.gaship.gashipshoppingmall.addresslist.exception.NotFoundAddressListException;
 import shop.gaship.gashipshoppingmall.addresslist.repository.AddressListRepository;
+import shop.gaship.gashipshoppingmall.addresslist.service.AddressListService;
 import shop.gaship.gashipshoppingmall.member.entity.Member;
 import shop.gaship.gashipshoppingmall.member.exception.MemberNotFoundException;
 import shop.gaship.gashipshoppingmall.member.repository.MemberRepository;
